backend/requirements.txt
django
djangorestframework
django-cors-headers
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv 